+++ src/modules/__mocks__/counter.js
